Qt4 supports 64bit types directly, so this Qt3 hack isn't needed any more. Only llightly tested, with HDHR-DVB.
git-svn-id: http://svn.mythtv.org/svn/trunk@20148 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/channelutil.cpp
@@ -729,7 +729,7 @@ bool ChannelUtil::GetTuningParams(uint      mplexid,
 
     dvb_transportid = query.value(0).toUInt();
     dvb_networkid   = query.value(1).toUInt();
-    frequency       = (uint64_t) query.value(2).toDouble(); // Qt 3.1 compat
+    frequency       = query.value(2).toUInt();
     modulation      = query.value(3).toString();
     si_std          = query.value(4).toString();
